[Downloader] 생성, ImageURL이 JSONDataFetcher대신 Downloader와 협력하도록 수정
