+++ StoreApp/StoreApp/ImageURL.swift
@@ -7,7 +7,6 @@
 //
 
 import Foundation
-import JSONDataFetcher
 
 struct ImageURL {
     let url: URL
@@ -16,7 +15,7 @@ struct ImageURL {
         guard let cacheURL = URL.cachesDirectory() else { return }
         let imagePath = cacheURL.appendingPathComponent(url.lastPathComponent)
         guard FileManager.default.fileExists(atPath: imagePath.path) == false else { return }
-        let jsonFetcher = JSONDataFetcher()
+        let downloader = Downloader()
         let downloadImage = { (fileURL: URL) -> Void in
             try? FileManager.default.moveItem(at: fileURL, to: imagePath)
             let userInfo = [UserInfoKey.imageName: self.url.lastPathComponent]
@@ -24,25 +23,6 @@ struct ImageURL {
                                             object: nil,
                                             userInfo: userInfo)
         }
-        jsonFetcher.download(with: url, successHandler: downloadImage)
-    }
-}
-
-extension JSONDataFetcher {
-    func download(with url: URL, successHandler: @escaping (URL) -> Void) {
-        URLSession.shared.downloadTask(with: url) { (fileURL, response, error) in
-            if let error = error {
-                print(error.localizedDescription)
-                return
-            }
-            
-            guard let fileURL = fileURL,
-                let response = response as? HTTPURLResponse,
-                (200...299) ~= response.statusCode,
-                let cacheURL = URL.cachesDirectory() else { return }
-            
-            successHandler(fileURL)
-            
-            }.resume()
+        downloader.download(with: url, successHandler: downloadImage)
     }
 }
